Refactor tests and update examples
