+++ examples/browser.html
@@ -20,7 +20,7 @@
               <input type="text" class="form-control" id="client-secret" placeholder="Client Secret" name="client-secret">
             </div>
             <div class="form-group">
-              <label for="custom-auth-url">Client Secret</label>
+              <label for="custom-auth-url">Custom Auth Url</label>
               <input type="text" class="form-control" id="custom-auth-url" placeholder="Custom Auth Url" name="custom-auth-url">
             </div>
             <div class="form-group">
